test(composer-app): phase 9 — dev-server startup harness
Adds a parallel harness for `vite serve` (the dev server) so dev-time
load improvements can be measured against the same BENCHMARKS ledger as
production preview. Refactors the existing prod-harness helpers
(trackNetwork, collectStartupReport, appendBenchmarkRow, …) into
harness-helpers.ts so the dev spec can reuse them.

New artifacts:
- src/playwright/harness-helpers.ts (extracted from startup.spec.ts)
- src/playwright/dev-startup.spec.ts (one `dev-cold` scenario: prime
  vite, then measure a fresh-context navigation against the warm dev
  server)
- src/playwright/playwright-dev.config.ts (webServer = `pnpm vite
  --port 4173`, testMatch scoped to dev-startup.spec.ts)
- moon task `composer-app:e2e-dev`

First baseline: profilerTotal=6,269 ms, navToReady=17,586 ms,
transferred=123 MB, modules=258. Compared to prod preview (~5,400 / ~8,700
/ 41 MB / 257), dev pays ~9 s of pre-main:start time on per-file
transform — exactly what phase 10 will target.
